:truck: move utils.js
